updated x509/verify.go to stadlib porting changes to x509/verify_modified.go
